remove autoconfig refactoring; add bean config condition
