+++ spring-cloud-gcp-autoconfigure/src/main/java/org/springframework/cloud/gcp/autoconfigure/pubsub/GcpPubSubAutoConfiguration.java
@@ -41,8 +41,6 @@ import org.threeten.bp.Duration;
 
 import org.springframework.beans.factory.ObjectProvider;
 import org.springframework.beans.factory.annotation.Qualifier;
-import org.springframework.boot.actuate.autoconfigure.health.ConditionalOnEnabledHealthIndicator;
-import org.springframework.boot.actuate.health.HealthIndicator;
 import org.springframework.boot.autoconfigure.AutoConfigureAfter;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnClass;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
@@ -57,7 +55,6 @@ import org.springframework.cloud.gcp.pubsub.core.PubSubException;
 import org.springframework.cloud.gcp.pubsub.core.PubSubTemplate;
 import org.springframework.cloud.gcp.pubsub.core.publisher.PubSubPublisherTemplate;
 import org.springframework.cloud.gcp.pubsub.core.subscriber.PubSubSubscriberTemplate;
-import org.springframework.cloud.gcp.pubsub.health.PubSubHealthIndicator;
 import org.springframework.cloud.gcp.pubsub.support.DefaultPublisherFactory;
 import org.springframework.cloud.gcp.pubsub.support.DefaultSubscriberFactory;
 import org.springframework.cloud.gcp.pubsub.support.PublisherFactory;
@@ -393,19 +390,4 @@ public class GcpPubSubAutoConfiguration {
 			.build();
 	}
 
-	/**
-	 * Spring Boot Actuator healthcheck autoconfiguration for Cloud Pub/Sub.
-	 */
-	@Configuration (proxyBeanMethods = false)
-	@ConditionalOnClass(HealthIndicator.class)
-	@ConditionalOnEnabledHealthIndicator("pubsub")
-	protected static class PubSubHealthIndicatorAutoConfiguration {
-
-		@Bean
-		@ConditionalOnMissingBean
-		public PubSubHealthIndicator pubSubHealthIndicator(PubSubTemplate pubSubTemplate) {
-			return new PubSubHealthIndicator(pubSubTemplate);
-		}
-
-	}
 }
