+++ spring-cloud-gcp-autoconfigure/src/main/java/org/springframework/cloud/gcp/autoconfigure/pubsub/health/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Auto-configuration for Spring Data Cloud Pub/Sub Health module.
+ */
+package org.springframework.cloud.gcp.autoconfigure.pubsub.health;
