+++ spring-cloud-gcp-autoconfigure/src/main/resources/META-INF/spring.factories
@@ -20,6 +20,7 @@ org.springframework.cloud.gcp.autoconfigure.datastore.GcpDatastoreEmulatorAutoCo
 org.springframework.cloud.gcp.autoconfigure.bigquery.GcpBigQueryAutoConfiguration,\
 org.springframework.cloud.gcp.autoconfigure.datastore.DatastoreTransactionManagerAutoConfiguration,\
 org.springframework.cloud.gcp.autoconfigure.firestore.FirestoreRepositoriesAutoConfiguration,\
+org.springframework.cloud.gcp.autoconfigure.pubsub.health.PubSubHealthIndicatorAutoConfiguration,\
 org.springframework.cloud.gcp.autoconfigure.secretmanager.GcpSecretManagerAutoConfiguration
 
 org.springframework.cloud.bootstrap.BootstrapConfiguration=\
